Ensure everything is using doubles

+++ src/main/java/info/cameronlund/autonplanner/actions/ClawAutonAction.java
@@ -176,7 +176,7 @@ public class ClawAutonAction extends AutonAction {
         g.setColor(getColor());
         Graphics2D g2 = (Graphics2D) g;
         g2.setFont(new Font("default", Font.BOLD, 14));
-        g2.drawString("C", robot.getPosX() - 6, robot.getPosY() + 5);
+        g2.drawString("C", (int) robot.getPosX() - 6, (int) robot.getPosY() + 5);
         return renderWithoutGraphics(robot);
     }
