+++ src/main/java/info/cameronlund/autonplanner/actions/WaitAutonAction.java
@@ -111,7 +111,7 @@ public class WaitAutonAction extends AutonAction {
         g.setColor(getColor());
         Graphics2D g2 = (Graphics2D) g;
         g2.setFont(new Font("default", Font.BOLD, 14));
-        g2.drawString("W", robot.getPosX() - 6, robot.getPosY() + 5);
+        g2.drawString("W",(int) robot.getPosX() - 6,(int) robot.getPosY() + 5);
         return renderWithoutGraphics(robot);
     }
